requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   google-genai
    #   httpx
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
google-auth==2.38.0
    # via google-genai
google-genai==1.5.0
    # via ankimate (pyproject.toml)
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via google-genai
pydantic-core==2.27.2
    # via pydantic
pypdf2==3.0.1
    # via ankimate (pyproject.toml)
requests==2.32.3
    # via google-genai
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
typing-extensions==4.12.2
    # via
    #   anyio
    #   google-genai
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websockets==14.2
    # via google-genai
